Add RawEmail#mail
Returns a `Mail` instance from the `RawEmail#data`.

I want to display some of the headers in the admin view of the
`RawEmail` but before this commit we'd have to parse the `RawEmail#data`
in the controller or something. Its better to just add this method here
and call the fields we want.

The specs are a _little_ heavy on mocks and doubles. I think they're
okay as its harder to test against a `Mail` instance, because the
message ID and date attributes change.